requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
phonenumbers==8.12.52
Pillow==9.2.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
